Loading physics setting from a bullet file
Bullet file are used to export physics settings from tools like Blender.
This file contains rigid body and constraints settings that can be
applied to scene objects.

A bullet file located in the 'assets' directory of an application can
now be loaded and the rigid bodies contained in this file will be
created and attached to scene objects. The scene objects must be created
first and have same names as the bodies in the bullet file.

GearVRf-DCO-1.0-Signed-off-by: Cristiano Bozzetto <c.bozzetto@samsung.com>