Changed stategy for adding constraints to world
A constraint is now added to the physics simulation world by the rigid
bodies linked by it. Both bodies will try to add it right after the
constraint is created or after the rigid bodies are added to the world.
GVRWorld ignores attemps to add a constraint that was already added.

Now the rigid bodies will keep refference of its constraints. This
approach allows easely removal of constraints from the world if at least
one of the linked bodies is removed.