allow picking of scene objects without meshes
It should be OK to add a sphere or box collider on a scene object with no meshes and have it use the bounds of the object and its children for the collider. This did not used to work because picking was done after culling. Now picking is done curing culling and
scene objects which are not culled are pickable even if they dont have meshes.